refactor component into smaller ones
